enable layer 3 PoDA service through standard bridge and cross domain messenger
We want the ability for PoDA lookups to be available on Rollux (L2) so we need a way for the precompile for PoDA to happen on L1 but then send a message to L2 such that the L2 function can only be called by the L1 bridge. Fortunately there is a service already for this and L1StandardBridge already speaks to L2StandardBridge to send messages from L1 to L2 for bridging related reasons (erc20's), we use the same solution to also check PoDA on L1 then send message to L2 which stores the results in a mapping.

The consuming layer 3 solution would simply lookup the podaMap mapping via keccak hash (hash of the data) in its state transition function.

The process would be as follows for layer 3:

1) send data to PoDA (on L1 utxo chain) and wait for confirm
2) call NEVM (L1) appendSequencerBatchToL2() on L1StandardBridge which will send message to L2, wait for confirm
3) Use your state transition function on Rollux (L2) to transition L3 state by looking up L2StandardBridge.podaMap to ensure DA exists

Aside from this we removed batchinbox.sol and put it inside l1standardbridge now, so the address used inside op-node for batchinbox now refers to the standard bridge instead of batch inbox contract, the node logic shouldn't need to change.